moved old community bloc to another file, re-added blocking ability on communities, and fixed loading indicator not being in the center
