+++ lib/community/bloc/community_event.dart
@@ -1,74 +1,24 @@
 part of 'community_bloc.dart';
 
-abstract class CommunityEvent extends Equatable {
+sealed class CommunityEvent extends Equatable {
   const CommunityEvent();
 
   @override
   List<Object> get props => [];
 }
 
-class GetCommunityPostsEvent extends CommunityEvent {
-  final bool reset;
-  final SortType? sortType;
-  final PostListingType? listingType;
-  final int? communityId;
-  final String? communityName;
-
-  const GetCommunityPostsEvent({this.reset = false, this.sortType, this.listingType, this.communityId, this.communityName});
-}
-
-class VotePostEvent extends CommunityEvent {
-  final int postId;
-  final VoteType score;
-
-  const VotePostEvent({required this.postId, required this.score});
-}
-
-class SavePostEvent extends CommunityEvent {
-  final int postId;
-  final bool save;
-
-  const SavePostEvent({required this.postId, required this.save});
-}
-
-class ForceRefreshEvent extends CommunityEvent {}
-
-class ChangeCommunitySubsciptionStatusEvent extends CommunityEvent {
+final class CommunityActionEvent extends CommunityEvent {
+  /// This is the community id to perform the action upon
   final int communityId;
-  final bool follow;
 
-  const ChangeCommunitySubsciptionStatusEvent({required this.communityId, required this.follow});
+  /// This indicates the relevant action to perform on the community
+  final CommunityAction communityAction;
+
+  /// This indicates the value to assign the action to. It is of type dynamic to allow for any type
+  /// TODO: Change the dynamic type to the correct type(s) if possible
+  final dynamic value;
+
+  const CommunityActionEvent({required this.communityId, required this.communityAction, this.value});
 }
 
-class CreatePostEvent extends CommunityEvent {
-  final String name;
-  final String body;
-  final String? url;
-  final bool nsfw;
-
-  const CreatePostEvent({required this.name, required this.body, this.url, this.nsfw = false});
-}
-
-class MarkPostAsReadEvent extends CommunityEvent {
-  final int postId;
-  final bool read;
-
-  const MarkPostAsReadEvent({required this.postId, required this.read});
-}
-
-class UpdatePostEvent extends CommunityEvent {
-  final PostViewMedia postViewMedia;
-
-  const UpdatePostEvent({required this.postViewMedia});
-}
-
-class BlockCommunityEvent extends CommunityEvent {
-  final int communityId;
-  final bool block;
-
-  const BlockCommunityEvent({required this.communityId, this.block = false});
-}
-
-class DismissReadEvent extends CommunityEvent {
-  const DismissReadEvent();
-}
+final class CommunityClearMessageEvent extends CommunityEvent {}
