+++ lib/community/widgets/community_sidebar.dart
@@ -14,9 +14,11 @@ import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 
 import 'package:thunder/account/bloc/account_bloc.dart';
 import 'package:thunder/community/bloc/community_bloc.dart';
+import 'package:thunder/community/enums/community_action.dart';
 import 'package:thunder/community/pages/create_post_page.dart';
 import 'package:thunder/core/auth/bloc/auth_bloc.dart';
 import 'package:thunder/core/enums/local_settings.dart';
+import 'package:thunder/core/singletons/lemmy_client.dart';
 import 'package:thunder/core/singletons/preferences.dart';
 import 'package:thunder/shared/common_markdown_body.dart';
 import 'package:thunder/shared/snackbar.dart';
@@ -52,126 +54,129 @@ class _CommunitySidebarState extends State<CommunitySidebar> {
 
     CommunityView communityView = widget.fullCommunityView.communityView;
 
-    return Container(
-      alignment: Alignment.centerRight,
-      child: Dismissible(
-        key: Key(communityView.community.id.toString()),
-        onUpdate: (DismissUpdateDetails details) => details.reached ? widget.onDismiss() : null,
-        direction: DismissDirection.startToEnd,
-        child: FractionallySizedBox(
-          widthFactor: 0.8,
-          alignment: FractionalOffset.centerRight,
-          child: Container(
-            color: theme.colorScheme.background,
-            alignment: Alignment.topRight,
-            child: Column(
-              children: [
-                AnimatedSwitcher(
-                  duration: const Duration(milliseconds: 100),
-                  transitionBuilder: (Widget child, Animation<double> animation) {
-                    return SizeTransition(
-                      sizeFactor: animation,
-                      child: FadeTransition(opacity: animation, child: child),
-                    );
-                  },
-                  child: communityView.blocked == false
-                      ? Padding(
-                          padding: const EdgeInsets.only(top: 10, left: 12, right: 12, bottom: 4),
-                          child: CommunityActions(isUserLoggedIn: isUserLoggedIn, fullCommunityView: widget.fullCommunityView),
-                        )
-                      : null,
-                ),
-                AnimatedSwitcher(
-                  duration: const Duration(milliseconds: 150),
-                  transitionBuilder: (Widget child, Animation<double> animation) {
-                    return SizeTransition(
-                      sizeFactor: animation,
-                      child: FadeTransition(opacity: animation, child: child),
-                    );
-                  },
-                  child: communityView.subscribed != SubscribedType.subscribed && communityView.subscribed != SubscribedType.pending
-                      ? BlockCommunityButton(communityView: communityView, isUserLoggedIn: isUserLoggedIn)
-                      : null,
-                ),
-                const SizedBox(height: 10.0),
-                const Divider(height: 1, thickness: 2),
-                Container(
-                  alignment: Alignment.topCenter,
-                  padding: const EdgeInsets.symmetric(horizontal: 12.0),
-                  height: MediaQuery.of(context).size.height - 200,
-                  child: ListView(
-                    shrinkWrap: true,
-                    children: [
-                      CommonMarkdownBody(body: communityView.community.description ?? ''),
-                      const SidebarSectionHeader(value: "Stats"),
-                      Padding(
-                        padding: const EdgeInsets.symmetric(horizontal: 8.0),
-                        child: CommunityStatsList(communityView: communityView),
-                      ),
-                      const SidebarSectionHeader(value: "Moderators"),
-                      Padding(
-                        padding: const EdgeInsets.symmetric(horizontal: 8.0),
-                        child: CommunityModeratorList(fullCommunityView: widget.fullCommunityView),
-                      ),
-                      Container(
-                        child: widget.fullCommunityView.site != null
-                            ? Column(
-                                children: [
-                                  const SidebarSectionHeader(value: "Host Instance"),
-                                  Padding(
-                                    padding: const EdgeInsets.symmetric(horizontal: 8.0),
-                                    child: Column(
-                                      children: [
-                                        Row(
-                                          children: [
-                                            CircleAvatar(
-                                              backgroundColor: widget.fullCommunityView.site?.icon != null ? Colors.transparent : theme.colorScheme.secondaryContainer,
-                                              foregroundImage: widget.fullCommunityView.site?.icon != null ? CachedNetworkImageProvider(widget.fullCommunityView.site!.icon!) : null,
-                                              maxRadius: 24,
-                                              child: widget.fullCommunityView.site?.icon == null
-                                                  ? Text(
-                                                      widget.fullCommunityView.moderators.first.moderator!.name[0].toUpperCase(),
-                                                      semanticsLabel: '',
-                                                      style: const TextStyle(fontWeight: FontWeight.bold, fontSize: 16),
-                                                    )
-                                                  : null,
-                                            ),
-                                            const SizedBox(width: 16.0),
-                                            Column(
-                                              mainAxisAlignment: MainAxisAlignment.start,
-                                              crossAxisAlignment: CrossAxisAlignment.start,
-                                              mainAxisSize: MainAxisSize.min,
-                                              children: [
-                                                Text(
-                                                  widget.fullCommunityView.site?.name ?? '',
-                                                  overflow: TextOverflow.ellipsis,
-                                                  maxLines: 1,
-                                                  style: theme.textTheme.headlineSmall?.copyWith(fontWeight: FontWeight.w600),
-                                                ),
-                                                Flexible(
-                                                  child: Text(
-                                                    widget.fullCommunityView.site?.description ?? '',
-                                                    style: theme.textTheme.bodyMedium,
-                                                  ),
-                                                ),
-                                              ],
-                                            ),
-                                          ],
-                                        ),
-                                        const Divider(),
-                                        CommonMarkdownBody(body: widget.fullCommunityView.site?.sidebar ?? ''),
-                                      ],
-                                    ),
-                                  ),
-                                ],
-                              )
-                            : null,
-                      ),
-                      const SizedBox(height: 256)
-                    ],
+    return BlocProvider<CommunityBloc>(
+      create: (context) => CommunityBloc(lemmyClient: LemmyClient.instance),
+      child: Container(
+        alignment: Alignment.centerRight,
+        child: Dismissible(
+          key: Key(communityView.community.id.toString()),
+          onUpdate: (DismissUpdateDetails details) => details.reached ? widget.onDismiss() : null,
+          direction: DismissDirection.startToEnd,
+          child: FractionallySizedBox(
+            widthFactor: 0.8,
+            alignment: FractionalOffset.centerRight,
+            child: Container(
+              color: theme.colorScheme.background,
+              alignment: Alignment.topRight,
+              child: Column(
+                children: [
+                  AnimatedSwitcher(
+                    duration: const Duration(milliseconds: 100),
+                    transitionBuilder: (Widget child, Animation<double> animation) {
+                      return SizeTransition(
+                        sizeFactor: animation,
+                        child: FadeTransition(opacity: animation, child: child),
+                      );
+                    },
+                    child: communityView.blocked == false
+                        ? Padding(
+                            padding: const EdgeInsets.only(top: 10, left: 12, right: 12, bottom: 4),
+                            child: CommunityActions(isUserLoggedIn: isUserLoggedIn, fullCommunityView: widget.fullCommunityView),
+                          )
+                        : null,
                   ),
-                )
-              ],
+                  AnimatedSwitcher(
+                    duration: const Duration(milliseconds: 150),
+                    transitionBuilder: (Widget child, Animation<double> animation) {
+                      return SizeTransition(
+                        sizeFactor: animation,
+                        child: FadeTransition(opacity: animation, child: child),
+                      );
+                    },
+                    child: communityView.subscribed != SubscribedType.subscribed && communityView.subscribed != SubscribedType.pending
+                        ? BlockCommunityButton(communityView: communityView, isUserLoggedIn: isUserLoggedIn)
+                        : null,
+                  ),
+                  const SizedBox(height: 10.0),
+                  const Divider(height: 1, thickness: 2),
+                  Container(
+                    alignment: Alignment.topCenter,
+                    padding: const EdgeInsets.symmetric(horizontal: 12.0),
+                    height: MediaQuery.of(context).size.height - 200,
+                    child: ListView(
+                      shrinkWrap: true,
+                      children: [
+                        CommonMarkdownBody(body: communityView.community.description ?? ''),
+                        const SidebarSectionHeader(value: "Stats"),
+                        Padding(
+                          padding: const EdgeInsets.symmetric(horizontal: 8.0),
+                          child: CommunityStatsList(communityView: communityView),
+                        ),
+                        const SidebarSectionHeader(value: "Moderators"),
+                        Padding(
+                          padding: const EdgeInsets.symmetric(horizontal: 8.0),
+                          child: CommunityModeratorList(fullCommunityView: widget.fullCommunityView),
+                        ),
+                        Container(
+                          child: widget.fullCommunityView.site != null
+                              ? Column(
+                                  children: [
+                                    const SidebarSectionHeader(value: "Host Instance"),
+                                    Padding(
+                                      padding: const EdgeInsets.symmetric(horizontal: 8.0),
+                                      child: Column(
+                                        children: [
+                                          Row(
+                                            children: [
+                                              CircleAvatar(
+                                                backgroundColor: widget.fullCommunityView.site?.icon != null ? Colors.transparent : theme.colorScheme.secondaryContainer,
+                                                foregroundImage: widget.fullCommunityView.site?.icon != null ? CachedNetworkImageProvider(widget.fullCommunityView.site!.icon!) : null,
+                                                maxRadius: 24,
+                                                child: widget.fullCommunityView.site?.icon == null
+                                                    ? Text(
+                                                        widget.fullCommunityView.moderators.first.moderator!.name[0].toUpperCase(),
+                                                        semanticsLabel: '',
+                                                        style: const TextStyle(fontWeight: FontWeight.bold, fontSize: 16),
+                                                      )
+                                                    : null,
+                                              ),
+                                              const SizedBox(width: 16.0),
+                                              Column(
+                                                mainAxisAlignment: MainAxisAlignment.start,
+                                                crossAxisAlignment: CrossAxisAlignment.start,
+                                                mainAxisSize: MainAxisSize.min,
+                                                children: [
+                                                  Text(
+                                                    widget.fullCommunityView.site?.name ?? '',
+                                                    overflow: TextOverflow.ellipsis,
+                                                    maxLines: 1,
+                                                    style: theme.textTheme.headlineSmall?.copyWith(fontWeight: FontWeight.w600),
+                                                  ),
+                                                  Flexible(
+                                                    child: Text(
+                                                      widget.fullCommunityView.site?.description ?? '',
+                                                      style: theme.textTheme.bodyMedium,
+                                                    ),
+                                                  ),
+                                                ],
+                                              ),
+                                            ],
+                                          ),
+                                          const Divider(),
+                                          CommonMarkdownBody(body: widget.fullCommunityView.site?.sidebar ?? ''),
+                                        ],
+                                      ),
+                                    ),
+                                  ],
+                                )
+                              : null,
+                        ),
+                        const SizedBox(height: 256)
+                      ],
+                    ),
+                  )
+                ],
+              ),
             ),
           ),
         ),
@@ -299,33 +304,45 @@ class BlockCommunityButton extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Padding(
-      padding: EdgeInsets.only(top: communityView.blocked ? 10 : 4, left: 12, right: 12, bottom: 4),
-      child: ElevatedButton(
-        onPressed: isUserLoggedIn
-            ? () {
-                HapticFeedback.heavyImpact();
-                hideSnackbar(context);
-                context.read<CommunityBloc>().add(BlockCommunityEvent(communityId: communityView.community.id, block: communityView.blocked == true ? false : true));
-              }
-            : null,
-        style: TextButton.styleFrom(
-          fixedSize: const Size.fromHeight(40),
-          foregroundColor: Colors.redAccent,
-          padding: EdgeInsets.zero,
-        ),
-        child: Row(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            Icon(
-              communityView.blocked == true ? Icons.undo_rounded : Icons.block_rounded,
-              semanticLabel: communityView.blocked == true ? 'Unblock Community' : 'Block Community',
+    return BlocBuilder<CommunityBloc, CommunityState>(
+      builder: (context, state) {
+        bool blocked = false;
+
+        if (state.communityView != null) {
+          blocked = state.communityView!.blocked;
+        } else {
+          blocked = communityView.blocked;
+        }
+
+        return Padding(
+          padding: EdgeInsets.only(top: blocked ? 10 : 4, left: 12, right: 12, bottom: 4),
+          child: ElevatedButton(
+            onPressed: isUserLoggedIn
+                ? () {
+                    HapticFeedback.heavyImpact();
+                    hideSnackbar(context);
+                    context.read<CommunityBloc>().add(CommunityActionEvent(communityAction: CommunityAction.block, communityId: communityView.community.id, value: !blocked));
+                  }
+                : null,
+            style: TextButton.styleFrom(
+              fixedSize: const Size.fromHeight(40),
+              foregroundColor: Colors.redAccent,
+              padding: EdgeInsets.zero,
             ),
-            const SizedBox(width: 4.0),
-            Text(communityView.blocked == true ? 'Unblock Community' : 'Block Community'),
-          ],
-        ),
-      ),
+            child: Row(
+              mainAxisAlignment: MainAxisAlignment.center,
+              children: [
+                Icon(
+                  blocked ? Icons.undo_rounded : Icons.block_rounded,
+                  semanticLabel: (state.communityView?.blocked ?? communityView.blocked) ? 'Unblock Community' : 'Block Community',
+                ),
+                const SizedBox(width: 4.0),
+                Text(blocked ? 'Unblock Community' : 'Block Community'),
+              ],
+            ),
+          ),
+        );
+      },
     );
   }
 }
@@ -425,12 +442,12 @@ class CommunityActions extends StatelessWidget {
             onPressed: isUserLoggedIn
                 ? () {
                     HapticFeedback.mediumImpact();
-                    context.read<CommunityBloc>().add(
-                          ChangeCommunitySubsciptionStatusEvent(
-                            communityId: communityView.community.id,
-                            follow: communityView.subscribed == SubscribedType.notSubscribed ? true : false,
-                          ),
-                        );
+                    // context.read<CommunityBloc>().add(
+                    //       ChangeCommunitySubsciptionStatusEvent(
+                    //         communityId: communityView.community.id,
+                    //         follow: communityView.subscribed == SubscribedType.notSubscribed ? true : false,
+                    //       ),
+                    //     );
                   }
                 : null,
             style: TextButton.styleFrom(
