+++ lib/community/widgets/post_card.dart
@@ -8,7 +8,7 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:swipeable_page_route/swipeable_page_route.dart';
 
 import 'package:thunder/account/bloc/account_bloc.dart';
-import 'package:thunder/community/bloc/community_bloc.dart';
+import 'package:thunder/community/bloc/community_bloc_old.dart';
 import 'package:thunder/community/utils/post_actions.dart';
 import 'package:thunder/community/utils/post_card_action_helpers.dart';
 import 'package:thunder/community/widgets/post_card_view_comfortable.dart';
@@ -245,7 +245,7 @@ class _PostCardState extends State<PostCard> {
                   PostCardAction.upvote,
                   PostCardAction.downvote,
                   PostCardAction.save,
-                  PostCardAction.toggelRead,
+                  PostCardAction.toggleRead,
                   PostCardAction.share,
                 ],
               ),
